std: pushing existing code in net.rs -> net_ip.rs and re-import/exporting
